# Restore safe casts and benchmarks from #2