@Daniel-B-Smith 2018-10-11
Thanks for the `benchstat` reference. I used benchcmp and found it mostly useless. Being able to run `-count` is what I really wanted.