@niamster 2018-11-19
@Daniel-B-Smith correct me if I'm wrong but your benchmark numbers are not accurate. You gained 500ns/op per 100 hash calls, which means that you've gained 5ns/op. I don't know if this is worth the risk if ABI changes. You should probably add an `init` function to the module that verifies your assumption so the issue could be detected early.
If I understand correctly the initial issue the old version of go compiler did a copy of string during the conversion but now it sees that the converted byte slice is not modified and thus does nothing