@Daniel-B-Smith 2018-11-19
If you look at the assembler, there is an extra call to https://golang.org/pkg/runtime/?m=all#stringtoslicebyte, which is where the performance difference is coming from.

This pattern is specifically endorsed by the Go documentation: https://golang.org/pkg/unsafe/#Pointer
Moreover, breaking this pattern would break many to most of the most important projects written in Go. The first three I looked at:

https://github.com/kubernetes/kubernetes/search?q=*%28*%5B%5Dbyte%29%28unsafe.Pointer&unscoped_q=*%28*%5B%5Dbyte%29%28unsafe.Pointer
https://github.com/etcd-io/etcd/search?q=*%28*%5B%5Dbyte%29%28unsafe.Pointer&unscoped_q=*%28*%5B%5Dbyte%29%28unsafe.Pointer
https://github.com/cockroachdb/cockroach/search?q=*%28*%5B%5Dbyte%29%28unsafe.Pointer&unscoped_q=*%28*%5B%5Dbyte%29%28unsafe.Pointer

This is a widely used library that I'm not going to make slower without a really good reason. If you want to add such an init function, feel free.