@Viq111 2018-11-19
@Daniel-B-Smith Should we backport to https://github.com/DataDog/mmh3 too ?